+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - REMIX-4582 + REMIX-4583: Added a new Additional Filters menu to the Stage Manager
 - REMIX-3539: Added "Delete Graph" Action Plugin Widget for the Stage Manager
 - REMIX-4757: Added the 1.3 Remix Logic component library.
+- REMIX-4778: Added logic graph create/edit functionality to properties panels and graph editor
+- REMIX-4739: Add support for specific icons for the different logic node categories
 
 ### Changed
 - Update hdremix and omni_core_materials to ext-822f7b6-main
@@ -32,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Optimized internal CI pipeline stages
 - Update hdremix and omni_core_materials to ext-2c10e95-main
 - REMIX-4721: Reimplemented Sidebar using ui.ToolBar for improved docking behavior
+- REMIX-4715: Make "Edit Graph" dialog a bit easier to use
 
 ### Fixed
 - Improved Pyright configuration by generating a list of search paths